Refactor: add two method to select page number and to set next, prev page value. Changes about data type.

+++ PaQuery/Enums/PageType.cs
@@ -1,6 +1,6 @@
 namespace PaQuery.Enums
 {
-    public enum PageType : int
+    public enum PageType : sbyte
     {
         Previous = -1,
         Next = 1
